Move `primary_key` package to `lib/rdms`
